Fixed bug when using a single path argument
The code was expecting an array of paths and this worked fine if the
command used something like:

spsave-watch --path ./a --path ./b

As yargs would convert that to an array. If only a single path was given,
the yargs value would be a string and any code expecting an array would
fail.

The mergeArgs() and mergePromptResults() functions have been changed to
ensure an array is always returned.